Fixed issue with BuyAofXGetBofYFZ whereby the incorrect result was given when:
(The selector and target Id were the same) AND (the DIFF of #selectors and #targets was > 1.)

Previously there was a special case to handle this, but it only handled the special case for 1 extra selector.
Now it handles the selectors and targets in a more even handed way.
It looks for both in the evluation at the same time mostly without special cases.

Also, removed inefficient counting O(n).